Rework all remaining shapes, add accurate outlines to other slopes, fix interactions with slope layers
