+++ common/src/main/java/com/copycatsplus/copycats/content/copycat/beam/CopycatBeamBlock.java
@@ -2,7 +2,6 @@ package com.copycatsplus.copycats.content.copycat.beam;
 
 import com.copycatsplus.copycats.CCBlocks;
 import com.copycatsplus.copycats.CCShapes;
-import com.copycatsplus.copycats.CCSimpleShapes;
 import com.copycatsplus.copycats.foundation.copycat.CCWaterloggedCopycatBlock;
 import com.copycatsplus.copycats.foundation.copycat.ICopycatBlock;
 import com.copycatsplus.copycats.foundation.copycat.IStateType;
@@ -129,7 +128,7 @@ public class CopycatBeamBlock extends CCWaterloggedCopycatBlock implements IStat
     @SuppressWarnings("deprecation")
     @Override
     public @NotNull VoxelShape getShape(BlockState pState, @NotNull BlockGetter pLevel, @NotNull BlockPos pPos, @NotNull CollisionContext pContext) {
-        return CCSimpleShapes.BEAM.get(pState.getValue(AXIS)).toShape();
+        return CCShapes.BEAM.get(pState.getValue(AXIS)).toShape();
     }
