+++ common/src/main/java/com/copycatsplus/copycats/content/copycat/board/CopycatBoardBlock.java
@@ -1,15 +1,12 @@
 package com.copycatsplus.copycats.content.copycat.board;
 
 import com.copycatsplus.copycats.CCShapes;
-import com.copycatsplus.copycats.CCSimpleShapes;
-import com.copycatsplus.copycats.foundation.copycat.ICopycatBlock;
 import com.copycatsplus.copycats.foundation.copycat.ICustomCTBlocking;
 import com.copycatsplus.copycats.foundation.copycat.multistate.IMultiStateCopycatBlock;
 import com.copycatsplus.copycats.foundation.copycat.multistate.IMultiStateCopycatBlockEntity;
 import com.copycatsplus.copycats.foundation.copycat.model.ScaledBlockAndTintGetter;
 import com.copycatsplus.copycats.foundation.copycat.multistate.WaterloggedMultiStateCopycatBlock;
 import com.google.common.collect.ImmutableMap;
-import com.simibubi.create.AllBlocks;
 import com.simibubi.create.content.contraptions.StructureTransform;
 import com.simibubi.create.foundation.utility.Iterate;
 import net.minecraft.MethodsReturnNonnullByDefault;
@@ -36,7 +33,6 @@ import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.block.state.properties.BooleanProperty;
 import net.minecraft.world.level.pathfinder.PathComputationType;
-import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.phys.shapes.BooleanOp;
 import net.minecraft.world.phys.shapes.CollisionContext;
@@ -161,7 +157,7 @@ public class CopycatBoardBlock extends WaterloggedMultiStateCopycatBlock impleme
         VoxelShape shape = Shapes.empty();
         for (Direction direction : Iterate.directions) {
             if (pState.getValue(byDirection(direction))) {
-                shape = Shapes.joinUnoptimized(shape, CCSimpleShapes.BOARD.get(direction).toShape(), BooleanOp.OR);
+                shape = Shapes.joinUnoptimized(shape, CCShapes.BOARD.get(direction).toShape(), BooleanOp.OR);
             }
         }
         return shape.optimize();
