+++ common/src/main/java/com/copycatsplus/copycats/content/copycat/slab/CopycatSlabBlock.java
@@ -2,14 +2,11 @@ package com.copycatsplus.copycats.content.copycat.slab;
 
 import com.copycatsplus.copycats.CCBlocks;
 import com.copycatsplus.copycats.CCShapes;
-import com.copycatsplus.copycats.CCSimpleShapes;
-import com.copycatsplus.copycats.foundation.copycat.ICopycatBlock;
 import com.copycatsplus.copycats.foundation.copycat.multistate.IMultiStateCopycatBlock;
 import com.copycatsplus.copycats.foundation.copycat.multistate.IMultiStateCopycatBlockEntity;
 import com.copycatsplus.copycats.foundation.copycat.model.ScaledBlockAndTintGetter;
 import com.copycatsplus.copycats.foundation.copycat.multistate.WaterloggedMultiStateCopycatBlock;
 import com.copycatsplus.copycats.utility.InteractionUtils;
-import com.simibubi.create.AllBlocks;
 import com.simibubi.create.content.contraptions.StructureTransform;
 import com.simibubi.create.foundation.placement.IPlacementHelper;
 import com.simibubi.create.foundation.placement.PlacementHelpers;
@@ -243,9 +240,9 @@ public class CopycatSlabBlock extends WaterloggedMultiStateCopycatBlock {
         if (type == SlabType.DOUBLE) {
             return Shapes.block();
         } else if (type == SlabType.BOTTOM) {
-            return CCSimpleShapes.SLAB_BOTTOM.get(axis).toShape();
+            return CCShapes.SLAB_BOTTOM.get(axis).toShape();
         } else {
-            return CCSimpleShapes.SLAB_TOP.get(axis).toShape();
+            return CCShapes.SLAB_TOP.get(axis).toShape();
         }
     }
